Refactor AbsentTests
