--- src/core-taggeds-optional/Optional.Tests/Absent.Static.Test/AbsentStaticTestStruct.cs
@@ -1,7 +1,0 @@
-﻿using PrimeFuncPack.UnitTest;
-
-namespace PrimeFuncPack.Core.Tests;
-
-public sealed class AbsentStaticTestStruct : AbsentStaticTestBase<StructType>
-{
-}
