--- src/core-taggeds-optional/Optional.Tests/Absent.Static.Test/AbsentStaticTestStructNullable.cs
@@ -1,7 +1,0 @@
-﻿using PrimeFuncPack.UnitTest;
-
-namespace PrimeFuncPack.Core.Tests;
-
-public sealed class AbsentStaticTestStructNullable : AbsentStaticTestBase<StructType?>
-{
-}
