--- src/core-taggeds-optional/Optional.Tests/Absent.Test/AbsentTestStruct.cs
@@ -1,7 +1,0 @@
-﻿using PrimeFuncPack.UnitTest;
-
-namespace PrimeFuncPack.Core.Tests;
-
-public sealed class AbsentTestStruct : AbsentTestBase<StructType>
-{
-}
