+++ src/core-taggeds-optional/Optional.Tests/AbsentStaticTests/AbsentStaticTestsRef.cs
@@ -2,6 +2,6 @@
 
 namespace PrimeFuncPack.Core.Tests;
 
-public sealed class AbsentTestRef : AbsentTestBase<RefType>
+public sealed class AbsentStaticTestsRef : AbsentStaticTestsBase<RefType>
 {
 }
