+++ src/core-taggeds-optional/Optional.Tests/AbsentStaticTests/AbsentStaticTestsStruct.cs
@@ -1,0 +1,7 @@
+﻿using PrimeFuncPack.UnitTest;
+
+namespace PrimeFuncPack.Core.Tests;
+
+public sealed class AbsentStaticTestsStruct : AbsentStaticTestsBase<StructType>
+{
+}
