+++ src/core-taggeds-optional/Optional.Tests/AbsentStaticTests/AbsentStaticTestsStructNullable.cs
@@ -1,0 +1,7 @@
+﻿using PrimeFuncPack.UnitTest;
+
+namespace PrimeFuncPack.Core.Tests;
+
+public sealed class AbsentStaticTestsStructNullable : AbsentStaticTestsBase<StructType?>
+{
+}
