+++ src/core-taggeds-optional/Optional.Tests/AbsentStaticTests/Base/AbsentStaticTestsBase.Factory.cs
@@ -1,0 +1,19 @@
+﻿using NUnit.Framework;
+using System;
+
+namespace PrimeFuncPack.Core.Tests;
+
+partial class AbsentStaticTestsBase<T>
+{
+    [Test]
+    public void Factory_Of_ExpectDefault()
+    {
+        Assert.AreEqual(default(Absent<T>), Absent.Of<T>());
+    }
+
+    [Test]
+    public void Factory_Of_ExpectNew()
+    {
+        Assert.AreEqual(new Absent<T>(), Absent.Of<T>());
+    }
+}
