+++ src/core-taggeds-optional/Optional.Tests/AbsentStaticTests/Base/AbsentStaticTestsBase.cs
@@ -1,0 +1,5 @@
+﻿namespace PrimeFuncPack.Core.Tests;
+
+public abstract partial class AbsentStaticTestsBase<T>
+{
+}
