+++ src/core-taggeds-optional/Optional.Tests/AbsentTests/AbsentTestsStructNullable.cs
@@ -1,0 +1,7 @@
+﻿using PrimeFuncPack.UnitTest;
+
+namespace PrimeFuncPack.Core.Tests;
+
+public sealed class AbsentTestsStructNullable : AbsentTestsBase<StructType?>
+{
+}
