+++ src/core-taggeds-optional/Optional.Tests/AbsentTests/Base/AbsentTestsBase.cs
@@ -1,0 +1,5 @@
+﻿namespace PrimeFuncPack.Core.Tests;
+
+public abstract partial class AbsentTestsBase<T>
+{
+}
